+++ src/service_provider/cmd/commands/MakeFileFSCommand.h
@@ -62,7 +62,7 @@ struct MakeFileFSCommand : public CommandBase {
 						m_terminal->write_ro(RODT_ATTR("Failed to create file: "));
 						m_terminal->write(filename);
 						m_terminal->write_ro(RODT_ATTR(" : "));
-						m_terminal->write(bStatus);
+						m_terminal->write((int32_t)bStatus);
 					}
 					delete[] filename;	
 				}
