+++ src/service_provider/cmd/commands/MoveFSCommand.h
@@ -76,7 +76,7 @@ struct MoveFSCommand : public CommandBase {
 						m_terminal->write_ro(RODT_ATTR(" : "));
 						m_terminal->write(newname);
 						m_terminal->write_ro(RODT_ATTR(" : "));
-						m_terminal->write(iStatus);
+						m_terminal->write((int32_t)iStatus);
 					}
 				}
